backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
feedparser==6.0.10
newspaper3k==0.2.8
playwright==1.40.0
sentence-transformers==2.2.2
spacy==3.5.4
psycopg2-binary==2.9.7
sqlalchemy==2.0.18
python-dotenv==1.0.0
faiss-cpu==1.7.4
pytest==7.4.0
pyyaml==6.0
requests==2.31.0
numpy==1.24.3
